requirements.txt
Flask
flask-cors
sentence-transformers
faiss-cpu
PyMuPDF
requests
